# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: THISCLOUD Daemon API
  description: |
    Central API for THISCLOUD hypervisor OS. Exposed by `thiscloudd` on port 8080.
    All resource endpoints are scoped to a tenant via `X-Tenant-Id` header (after T0.4).
  version: "1.0.0"
  contact:
    name: THISJOWI
servers:
  - url: http://127.0.0.1:8080
    description: Local daemon
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Authenticate user and get JWT
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Missing username or password
        "401":
          description: Invalid credentials

  # ── VMs ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from POST /auth/login. Roles:
        - admin: full access (wheel group members)
        - operator: same as admin, cannot modify auth settings
        - tenant_admin: full access within tenant
        - tenant_user: read-only + create own resources
        - auditor: read-only
...
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          maxLength: 32
          pattern: "^[a-zA-Z0-9._-]+$"
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT Bearer token
        role:
          type: string
          enum: [admin, tenant_user]
        tenant_id:
          type: string
        expires_at:
          type: integer
          description: Unix timestamp
...